# dash-backend/rest_api/swagger/index.yaml
definitions:
...
  VehicleAttributeSchema:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      description:
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      modification_date:
        format: date-time
        readOnly: true
        type: string
      name:
        type: string
      parent:
        allOf:
        - $ref: '#/definitions/VehicleAttributeSchema'
        readOnly: true
      parent_id:
        format: int32
        type: integer
    required:
    - description
    - name
    type: object
  VehicleAttributeValueSchema:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      modification_date:
        format: date-time
        readOnly: true
        type: string
      value:
        type: string
      value_in_number:
        format: float
        type: number
      vehicle_attribute:
        allOf:
        - $ref: '#/definitions/VehicleAttributeSchema'
        readOnly: true
      vehicle_attribute_id:
        format: int32
        type: integer
    required:
    - value
    type: object
...
  VehicleBrand:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
...
  VehicleCategory:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      description:
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
...
  VehicleDriveTrain:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      description:
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
  VehicleEngine:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      horse_power:
        format: int32
        type: integer
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      label:
        type: string
      modification_date:
        format: date-time
        readOnly: true
        type: string
      torque:
        format: int32
        type: integer
      type:
        enum:
        - OTHERS
        - GAS
        - DIESEL
        - ELECTRIC
        type: string
    required:
    - type
    type: object
  VehicleGalleryItemSchema:
    properties:
      asset_path:
        type: string
      creation_date:
        format: date-time
        readOnly: true
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      modification_date:
        format: date-time
        readOnly: true
        type: string
      order:
        format: int32
        type: integer
      type:
        enum:
        - COVER
        - INTERIOR
        - INTERIOR_COLOR
        - EXTERIOR
        - EXTERIOR_COLOR
        - BRAND_LOGO
        - MODEL_AVATAR
        - YEAR_MODEL_AVATAR
        - VIDEO_INTRODUCTION
        - MODEL_STYLE_AVATAR
        - OPTION_ITEM_AVATAR
        - VEHICLE_AVATAR
        - OTHERS
        type: string
    required:
    - asset_path
    - type
    type: object
...
  VehicleInfo:
    properties:
      city_mpg:
        format: int32
        type: integer
      combined_mpg:
        format: int32
        type: integer
      creation_date:
        format: date-time
        readOnly: true
        type: string
      description:
        type: string
      doors:
        format: int32
        type: integer
      highway_mpg:
        format: int32
        type: integer
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      safety_rate:
        format: int32
        type: integer
      seats:
        format: int32
        type: integer
      vehicle_warranty:
        $ref: '#/definitions/VehicleWarranty'
    type: object
  VehicleModel:
    properties:
      description:
        type: string
      name:
        type: string
      vehicle_brand:
        allOf:
        - $ref: '#/definitions/VehicleBrand'
        readOnly: true
    required:
    - description
    - name
    type: object
...
  VehicleTransmission:
    properties:
      creation_date:
        format: date-time
        readOnly: true
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      label:
        type: string
      modification_date:
        format: date-time
        readOnly: true
        type: string
      speeds:
        format: int32
        type: integer
      type:
        enum:
        - OTHERS
        - AUTOMATIC
        - MANUAL
        type: string
    required:
    - type
    type: object
  VehicleWarranty:
    properties:
      anti_corrosion:
        type: string
      comprehensive:
        type: string
      creation_date:
        format: date-time
        readOnly: true
        type: string
      id:
        format: int32
        type: integer
        x-nullable: true
      is_active:
        type: boolean
      modification_date:
        format: date-time
        readOnly: true
        type: string
      paint:
        type: string
      power_train:
        type: string
      roadside_assistance:
        type: string
    type: object
...
  VehicleYearModelSchema:
    properties:
      attributes:
        items:
          $ref: '#/definitions/VehicleAttributeValueSchema'
        readOnly: true
        type: array
      categories:
        items:
          $ref: '#/definitions/VehicleCategory'
        type: array
      creation_date:
        format: date-time
        readOnly: true
        type: string
      description:
        type: string
      drive_trains:
        items:
          $ref: '#/definitions/VehicleDriveTrain'
        type: array
      engines:
        items:
          $ref: '#/definitions/VehicleEngine'
        type: array
      galleries:
        items:
          $ref: '#/definitions/VehicleGalleryItemSchema'
        type: array
      id:
        format: int32
        type: integer
        x-nullable: true
      infos:
        items:
          $ref: '#/definitions/VehicleInfo'
        type: array
      modification_date:
        format: date-time
        readOnly: true
        type: string
      name:
        type: string
      transmissions:
        items:
          $ref: '#/definitions/VehicleTransmission'
        type: array
      vehicle_model:
        allOf:
        - $ref: '#/definitions/VehicleModel'
        readOnly: true
      vehicle_model_id:
        format: int32
        type: integer
      year:
        format: int32
        type: integer
    required:
    - description
    - name
    type: object
info:
  title: Swagger Petstore
  version: 1.0.0
paths:
...
  /vehicle_models/{vehicle_model_id}/vehicle_year_models:
    get:
      parameters:
      - in: query
        name: direction
        required: 'true'
        schema:
          default: asc
          enum:
          - asc
          - desc
          type: string
      - in: query
        name: is_active
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      - in: query
        name: limit
        required: 'true'
        schema:
          default: '100'
          type: integer
      - in: query
        name: offset
        required: 'true'
        schema:
          default: '0'
          type: integer
      - default: ''
        in: query
        name: q
        required: 'true'
        schema:
          type: string
      - in: query
        name: sort
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema: VehicleYearModelSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
    parameters:
    - description: vehicleModelId
      in: path
      name: vehicle_model_id
      required: true
...
  /vehicle_year_models:
    get:
      parameters:
      - in: query
        name: direction
        required: 'true'
        schema:
          default: asc
          enum:
          - asc
          - desc
          type: string
      - in: query
        name: is_active
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      - in: query
        name: limit
        required: 'true'
        schema:
          default: '100'
          type: integer
      - in: query
        name: offset
        required: 'true'
        schema:
          default: '0'
          type: integer
      - default: ''
        in: query
        name: q
        required: 'true'
        schema:
          type: string
      - in: query
        name: sort
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema: VehicleYearModelSchema
        '400': &id034
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401': &id035
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX: &id036
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
    post:
      parameters:
      - in: body
        name: VehicleYearModelBody
        required: 'true'
        schema:
          $ref: '#/definitions/VehicleYearModelSchema'
      responses:
        '200':
          content:
            application/json:
              schema: VehicleYearModelSchema
        '400': *id034
        '401': *id035
        5XX: *id036
  /vehicle_year_models/{vehicle_year_model_id}:
    delete:
      responses:
        '200':
          content:
            application/json:
              schema: VehicleYearModelSchema
        '400': &id037
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401': &id038
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX: &id039
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
    get:
      responses:
        '200':
          content:
            application/json:
              schema: VehicleYearModelSchema
        '400': *id037
        '401': *id038
        5XX: *id039
    parameters:
    - description: vehicleYearModelId
      in: path
      name: vehicle_year_model_id
      required: true
    put:
      parameters:
      - in: body
        name: VehicleYearModelBody
        required: 'true'
        schema:
          $ref: '#/definitions/VehicleYearModelSchema'
      responses:
        '200':
          content:
            application/json:
              schema: VehicleYearModelSchema
        '400': *id037
        '401': *id038
        5XX: *id039
  /vehicle_year_models/{vehicle_year_model_id}/attribute_values:
    post:
      parameters:
      - description: VehicleYearModelId
        in: path
        name: vehicle_year_model_id
        required: true
      - in: body
        name: VehicleAttributeValuesBody
        required: 'true'
        schema:
          $ref: '#/definitions/VehicleAttributeValueSchema'
      responses:
        '200':
          content:
            application/json:
              schema: VehicleYearModelSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
  /vehicle_year_models/{vehicle_year_model_id}/avatar:
    post:
      parameters:
      - description: VehicleYearModelId
        in: path
        name: vehicle_year_model_id
        required: true
      - description: VehicleYearModelAvatar
        in: formData
        name: VehicleYearModelAvatar
        required: 'true'
        type: file
      responses:
        '200':
          content:
            application/json:
              schema: VehicleYearModelSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
  /vehicle_year_models/{vehicle_year_model_id}/cover:
    post:
      parameters:
      - description: VehicleYearModelId
        in: path
        name: vehicle_year_model_id
        required: true
      - description: VehicleYearModelCover
        in: formData
        name: VehicleYearModelCover
        required: 'true'
        type: file
      responses:
        '200':
          content:
            application/json:
              schema: VehicleYearModelSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
  /vehicle_year_models/{vehicle_year_model_id}/exterior:
    post:
      parameters:
      - description: VehicleYearModelId
        in: path
        name: vehicle_year_model_id
        required: true
      - description: VehicleYearModelExterior
        in: formData
        name: VehicleYearModelExterior
        required: 'true'
        type: file
      - description: VehicleYearModelExterior2
        in: formData
        name: VehicleYearModelExterior2
        required: 'true'
        type: file
      responses:
        '200':
          content:
            application/json:
              schema: VehicleYearModelSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
  /vehicle_year_models/{vehicle_year_model_id}/interior:
    post:
      parameters:
      - description: VehicleYearModelId
        in: path
        name: vehicle_year_model_id
        required: true
      - description: VehicleYearModelInterior
        in: formData
        name: VehicleYearModelInterior
        required: 'true'
        type: file
      - description: VehicleYearModelInterior2
        in: formData
        name: VehicleYearModelInterior2
        required: 'true'
        type: file
      responses:
        '200':
          content:
            application/json:
              schema: VehicleYearModelSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
  /vehicle_year_models/{vehicle_year_model_id}/introduction_video:
    post:
      parameters:
      - description: VehicleYearModelId
        in: path
        name: vehicle_year_model_id
        required: true
      - description: VehicleYearModelIntroductionVideo
        in: formData
        name: VehicleYearModelIntroductionVideo
        required: 'true'
        type: file
      responses:
        '200':
          content:
            application/json:
              schema: VehicleYearModelSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
  /vehicle_year_models/{vehicle_year_model_id}/vehicle_model_styles:
    get:
      parameters:
      - in: query
        name: direction
        required: 'true'
        schema:
          default: asc
          enum:
          - asc
          - desc
          type: string
      - in: query
        name: is_active
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      - in: query
        name: limit
        required: 'true'
        schema:
          default: '100'
          type: integer
      - in: query
        name: offset
        required: 'true'
        schema:
          default: '0'
          type: integer
      - default: ''
        in: query
        name: q
        required: 'true'
        schema:
          type: string
      - in: query
        name: sort
        required: 'true'
        schema:
          default: 'true'
          type: boolean
      responses:
        '200':
          content:
            application/json:
              schema: VehicleModelStyleSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
    parameters:
    - description: vehicleYearModelId
      in: path
      name: vehicle_year_model_id
      required: true
...
  /vehicles/{vehicle_id}:
    get:
      parameters:
      - description: VehicleId
        in: path
        name: vehicle_id
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: VehicleSchema
        '400':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Bad Requests
        '401':
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Not Authorized
        5XX:
          content:
            application/json:
              schema: ExceptionResponseSchema
          description: Error From Server
...
swagger: 2.0.0